+++ internal/webhook/inferenceservice/register.go
@@ -1,6 +1,6 @@
 //go:build !nowebhook
 
-package dataconnection
+package inferenceservice
 
 import (
 	ctrl "sigs.k8s.io/controller-runtime"
