+++ internal/webhook/webhook.go
@@ -6,10 +6,10 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 
 	authwebhook "github.com/opendatahub-io/opendatahub-operator/v2/internal/webhook/auth"
-	dc "github.com/opendatahub-io/opendatahub-operator/v2/internal/webhook/dataconnection"
 	dscwebhook "github.com/opendatahub-io/opendatahub-operator/v2/internal/webhook/datasciencecluster"
 	dsciwebhook "github.com/opendatahub-io/opendatahub-operator/v2/internal/webhook/dscinitialization"
 	hardwareprofilewebhook "github.com/opendatahub-io/opendatahub-operator/v2/internal/webhook/hardwareprofile"
+	isvc "github.com/opendatahub-io/opendatahub-operator/v2/internal/webhook/inferenceservice"
 	kueuewebhook "github.com/opendatahub-io/opendatahub-operator/v2/internal/webhook/kueue"
 )
 
@@ -22,7 +22,7 @@ func RegisterAllWebhooks(mgr ctrl.Manager) error {
 		authwebhook.RegisterWebhooks,
 		hardwareprofilewebhook.RegisterWebhooks,
 		kueuewebhook.RegisterWebhooks,
-		dc.RegisterWebhooks,
+		isvc.RegisterWebhooks,
 	}
 	for _, reg := range webhookRegistrations {
 		if err := reg(mgr); err != nil {
